Update to ubuntu 24.04 containers, fix test extension
Update the dev containers to use current Ubuntu 24.04
LTS release.  This should be what github actions
will be using in a bit in any case, we should
have no problem when they switch.

Also the catch2 extension we were using seems to have
just stopped working.  The VSCode testing ui is changing,
so we will need to look into this.  But for now we
change to this forked extension of the original
test runner which seems to be working for now as
we are expecting.