Undefine task 1 and fix description error
There was an error in the assignment description just before
beginning task 1 indicating that no changes would be
made in the header file.  Removed this bad instruction
as current version does require sutdents to declare the
member functions in the header file.

Also the first task1 had been incorrectly already defined,
change it back so it is initially undefined.